Dead code removal.  move test-only code to tmol.test
