--- scripts/ligand_prep/designs.prot.smi
@@ -1,2 +1,0 @@
-Cc1cn([C@@H]2C[C@@H](O)[C@H](CO)O2)c(=O)[nH]c1=O	ref1
-Cc1cc2nc3c(=O)[nH]c(=O)nc-3n(C[C@H](O)[C@H](O)[C@H](O)CO)c2cc1C	ref2
