--- scripts/ligand_prep/legacy/prepare_smiles.sh
@@ -1,3 +1,0 @@
-app=/home/gzhou/git/dimorphite_dl/dimorphite_dl.py
-
-python $app --smiles_file ./designs.smi --min_ph 7.4 --max_ph 7.4 --output_file designs.prot.smi --pka_precision 0.1
