--- scripts/ligand_prep/params/ref1.params
@@ -1,108 +1,0 @@
-NAME ref1
-IO_STRING LG1 Z
-TYPE LIGAND
-AA UNK
-ATOM C1   CS3  X     0.138
-ATOM C2   CR   X    -0.124
-ATOM C3   CR   X    -0.041
-ATOM N1   Nad3 X    -0.469
-ATOM C4   CS1  X     0.580
-ATOM C5   CS2  X     0.000
-ATOM C6   CSp  X     0.280
-ATOM O1   Ohx  X    -0.680
-ATOM C7   CS1  X     0.280
-ATOM C8   CSp  X     0.280
-ATOM O2   Ohx  X    -0.680
-ATOM O3   Oet  X    -0.560
-ATOM C9   CRp  X     0.690
-ATOM O4   Oad  X    -0.570
-ATOM N2   Nin  X    -0.490
-ATOM C10  CRp  X     0.616
-ATOM O5   Oad  X    -0.570
-ATOM HC1  HC   X     0.000
-ATOM HC2  HC   X     0.000
-ATOM HC3  HC   X     0.000
-ATOM HC4  HR   X     0.150
-ATOM HC5  HC   X     0.000
-ATOM HC6  HC   X     0.000
-ATOM HC7  HC   X     0.000
-ATOM HC8  HC   X     0.000
-ATOM HO1  HO   X     0.400
-ATOM HC9  HC   X     0.000
-ATOM HC10 HC   X     0.000
-ATOM HC11 HC   X     0.000
-ATOM HO2  HO   X     0.400
-ATOM HN1  HN   X     0.370
-BOND_TYPE C1   C2   1
-BOND_TYPE N1   C4   1
-BOND_TYPE C4   C5   1 RING
-BOND_TYPE C4   O3   1 RING
-BOND_TYPE C5   C6   1 RING
-BOND_TYPE C6   O1   1
-BOND_TYPE C6   C7   1 RING
-BOND_TYPE C7   C8   1
-BOND_TYPE C7   O3   1 RING
-BOND_TYPE C8   O2   1
-BOND_TYPE C1   HC1  1
-BOND_TYPE C1   HC2  1
-BOND_TYPE C1   HC3  1
-BOND_TYPE C3   HC4  1
-BOND_TYPE C4   HC5  1
-BOND_TYPE C5   HC6  1
-BOND_TYPE C5   HC7  1
-BOND_TYPE C6   HC8  1
-BOND_TYPE O1   HO1  1
-BOND_TYPE C7   HC9  1
-BOND_TYPE C8   HC10 1
-BOND_TYPE C8   HC11 1
-BOND_TYPE O2   HO2  1
-BOND_TYPE N2   HN1  1
-BOND_TYPE C9   O4   2
-BOND_TYPE C10  O5   2
-BOND_TYPE C2   C3   4 RING
-BOND_TYPE C2   C10  4 RING
-BOND_TYPE C3   N1   4 RING
-BOND_TYPE N1   C9   4 RING
-BOND_TYPE C9   N2   4 RING
-BOND_TYPE N2   C10  4 RING
-CUT_BOND   C7   O3
-CUT_BOND   N2  C10
-NBR_ATOM N1
-NBR_RADIUS 12.46368
-ICOOR_INTERNAL  N1      0.000000    0.000000    0.000000 N1   C3   C4  
-ICOOR_INTERNAL  C3      0.000000    0.000000    1.381496 N1   C3   C4  
-ICOOR_INTERNAL  C4      0.000000   58.139720    1.457142 N1   C3   C4  
-ICOOR_INTERNAL  C9    176.229111   58.567411    1.389279 N1   C3   C4  
-ICOOR_INTERNAL  C5     95.903755   66.874072    1.523814 C4   N1   C3  
-ICOOR_INTERNAL  O3    -24.906275   68.457062    1.442896 C4   N1   C3  
-ICOOR_INTERNAL  O4   -177.105865   56.104689    1.229135 C9   N1   C3  
-ICOOR_INTERNAL  N2      2.321273   64.220952    1.368017 C9   N1   C3  
-ICOOR_INTERNAL  C2   -177.593916   57.389101    1.338754 C3   N1   C4  
-ICOOR_INTERNAL  C6   -153.315122   77.805002    1.509344 C5   C4   N1  
-ICOOR_INTERNAL  C1   -179.841771   55.567006    1.493295 C2   C3   N1  
-ICOOR_INTERNAL  C10     0.292243   61.954967    1.481095 C2   C3   N1  
-ICOOR_INTERNAL  O1    -83.534110   70.566671    1.428715 C6   C5   C4  
-ICOOR_INTERNAL  C7     34.838714   76.998819    1.524248 C6   C5   C4  
-ICOOR_INTERNAL  O5    179.189543   57.974321    1.225833 C10  C2   C3  
-ICOOR_INTERNAL  C8     89.282240   67.054220    1.530706 C7   C6   C5  
-ICOOR_INTERNAL  O2   -178.627256   70.242514    1.427586 C8   C7   C6  
-ICOOR_INTERNAL  HC5   122.370078   71.299822    1.097553 C4   N1   C5  
-ICOOR_INTERNAL  HC4  -179.793177   62.910902    1.085507 C3   N1   C2  
-ICOOR_INTERNAL  HC6  -122.268745   68.192147    1.094310 C5   C4   C6  
-ICOOR_INTERNAL  HC7   117.003467   69.421699    1.096615 C5   C4   C6  
-ICOOR_INTERNAL  HN1  -177.489282   63.242338    1.007744 N2   C9   C10 
-ICOOR_INTERNAL  HC8  -119.033451   67.179206    1.093502 C6   C5   O1  
-ICOOR_INTERNAL  HC1   120.488238   69.460895    1.094920 C1   C2   C3  
-ICOOR_INTERNAL  HC2     0.617895   68.237200    1.095152 C1   C2   C3  
-ICOOR_INTERNAL  HC3  -119.278178   69.475564    1.094833 C1   C2   C3  
-ICOOR_INTERNAL  HO1    75.048369   70.394509    0.974041 O1   C6   C5  
-ICOOR_INTERNAL  HC9   125.344557   67.510748    1.095475 C7   C6   C8  
-ICOOR_INTERNAL  HC10  118.623417   68.311634    1.094212 C8   C7   O2  
-ICOOR_INTERNAL  HC11 -119.636880   68.797415    1.093878 C8   C7   O2  
-ICOOR_INTERNAL  HO2    38.677258   73.482388    0.976613 O2   C8   C7  
-CHI   1   C3   N1   C4   C5 #biaryl
-CHI   2   C6   C7   C8   O2 
-CHI   3   C5   C6   O1  HO1 
-PROTON_CHI 3 SAMPLES 3 60 -60 180 EXTRA 1 20
-CHI   4   C7   C8   O2  HO2 
-PROTON_CHI 4 SAMPLES 3 60 -60 180 EXTRA 1 20
