--- scripts/ligand_prep/params/ref2.params
@@ -1,165 +1,0 @@
-NAME ref2
-IO_STRING LG1 Z
-TYPE LIGAND
-AA UNK
-ATOM C1   CS3  X     0.143
-ATOM C2   CR   X    -0.143
-ATOM C3   CR   X    -0.150
-ATOM C4   CR   X     0.179
-ATOM N1   Nim  X    -0.629
-ATOM C5   CR   X     0.450
-ATOM C6   CRp  X     0.630
-ATOM O1   Oad  X    -0.570
-ATOM N2   Nin  X    -0.490
-ATOM C7   CRp  X     0.841
-ATOM O2   Oad  X    -0.570
-ATOM N3   Nim  X    -0.661
-ATOM C8   CR   X     0.500
-ATOM N4   Nad3 X    -0.519
-ATOM C9   CS2  X     0.369
-ATOM C10  CSp  X     0.280
-ATOM O3   Ohx  X    -0.680
-ATOM C11  CSp  X     0.280
-ATOM O4   Ohx  X    -0.680
-ATOM C12  CSp  X     0.280
-ATOM O5   Ohx  X    -0.680
-ATOM C13  CSp  X     0.280
-ATOM O6   Ohx  X    -0.680
-ATOM C14  CR   X     0.100
-ATOM C15  CR   X    -0.150
-ATOM C16  CR   X    -0.143
-ATOM C17  CS3  X     0.143
-ATOM HC1  HC   X     0.000
-ATOM HC2  HC   X     0.000
-ATOM HC3  HC   X     0.000
-ATOM HC4  HR   X     0.150
-ATOM HN1  HN   X     0.370
-ATOM HC5  HC   X     0.000
-ATOM HC6  HC   X     0.000
-ATOM HC7  HC   X     0.000
-ATOM HO1  HO   X     0.400
-ATOM HC8  HC   X     0.000
-ATOM HO2  HO   X     0.400
-ATOM HC9  HC   X     0.000
-ATOM HO3  HO   X     0.400
-ATOM HC10 HC   X     0.000
-ATOM HC11 HC   X     0.000
-ATOM HO4  HO   X     0.400
-ATOM HC12 HR   X     0.150
-ATOM HC13 HC   X     0.000
-ATOM HC14 HC   X     0.000
-ATOM HC15 HC   X     0.000
-BOND_TYPE C1   C2   1
-BOND_TYPE C5   C8   2 RING
-BOND_TYPE N4   C9   1
-BOND_TYPE C9   C10  1
-BOND_TYPE C10  O3   1
-BOND_TYPE C10  C11  1
-BOND_TYPE C11  O4   1
-BOND_TYPE C11  C12  1
-BOND_TYPE C12  O5   1
-BOND_TYPE C12  C13  1
-BOND_TYPE C13  O6   1
-BOND_TYPE C16  C17  1
-BOND_TYPE C1   HC1  1
-BOND_TYPE C1   HC2  1
-BOND_TYPE C1   HC3  1
-BOND_TYPE C3   HC4  1
-BOND_TYPE N2   HN1  1
-BOND_TYPE C9   HC5  1
-BOND_TYPE C9   HC6  1
-BOND_TYPE C10  HC7  1
-BOND_TYPE O3   HO1  1
-BOND_TYPE C11  HC8  1
-BOND_TYPE O4   HO2  1
-BOND_TYPE C12  HC9  1
-BOND_TYPE O5   HO3  1
-BOND_TYPE C13  HC10 1
-BOND_TYPE C13  HC11 1
-BOND_TYPE O6   HO4  1
-BOND_TYPE C15  HC12 1
-BOND_TYPE C17  HC13 1
-BOND_TYPE C17  HC14 1
-BOND_TYPE C17  HC15 1
-BOND_TYPE C6   O1   2
-BOND_TYPE C7   O2   2
-BOND_TYPE C2   C3   4 RING
-BOND_TYPE C2   C16  4 RING
-BOND_TYPE C3   C4   4 RING
-BOND_TYPE C4   N1   4 RING
-BOND_TYPE C4   C14  4 RING
-BOND_TYPE N1   C5   4 RING
-BOND_TYPE C5   C6   4 RING
-BOND_TYPE C6   N2   4 RING
-BOND_TYPE N2   C7   4 RING
-BOND_TYPE C7   N3   4 RING
-BOND_TYPE N3   C8   4 RING
-BOND_TYPE C8   N4   4 RING
-BOND_TYPE N4   C14  4 RING
-BOND_TYPE C14  C15  4 RING
-BOND_TYPE C15  C16  4 RING
-CUT_BOND   N3   C8
-CUT_BOND   N4  C14
-CUT_BOND  C15  C16
-NBR_ATOM C9
-NBR_RADIUS 16.34200
-ICOOR_INTERNAL  C9      0.000000    0.000000    0.000000 C9   N4   C10 
-ICOOR_INTERNAL  N4      0.000000    0.000000    1.476560 C9   N4   C10 
-ICOOR_INTERNAL  C10     0.000000   69.177668    1.545433 C9   N4   C10 
-ICOOR_INTERNAL  O3    -64.798442   69.325388    1.436111 C10  C9   N4  
-ICOOR_INTERNAL  C11   173.011475   69.187696    1.547376 C10  C9   N4  
-ICOOR_INTERNAL  C8    -81.328626   60.859675    1.392825 N4   C9   C10 
-ICOOR_INTERNAL  O4    -59.780797   70.262199    1.428822 C11  C10  C9  
-ICOOR_INTERNAL  C12   176.246079   67.876466    1.545283 C11  C10  C9  
-ICOOR_INTERNAL  C5   -173.442032   61.591298    1.508318 C8   N4   C9  
-ICOOR_INTERNAL  O5    -64.111640   68.135713    1.437672 C12  C11  C10 
-ICOOR_INTERNAL  C13   175.744026   67.870366    1.532987 C12  C11  C10 
-ICOOR_INTERNAL  N1     -2.487317   57.936605    1.297324 C5   C8   N4  
-ICOOR_INTERNAL  C6    177.258029   62.148374    1.504317 C5   C8   N4  
-ICOOR_INTERNAL  O6    178.901469   70.938092    1.429628 C13  C12  C11 
-ICOOR_INTERNAL  O1    169.746250   57.173186    1.219368 C6   C5   C8  
-ICOOR_INTERNAL  N2     -6.906445   66.694950    1.361326 C6   C5   C8  
-ICOOR_INTERNAL  C4     -0.160368   61.463073    1.407034 N1   C5   C8  
-ICOOR_INTERNAL  C7     21.533842   56.421555    1.364350 N2   C6   C5  
-ICOOR_INTERNAL  C3   -178.216806   63.476439    1.399810 C4   N1   C5  
-ICOOR_INTERNAL  C14     1.489985   56.975876    1.415125 C4   N1   C5  
-ICOOR_INTERNAL  O2    152.604584   61.376975    1.226069 C7   N2   C6  
-ICOOR_INTERNAL  N3    -26.140784   60.175944    1.366496 C7   N2   C6  
-ICOOR_INTERNAL  C15  -178.897292   62.652092    1.406036 C14  C4   N1  
-ICOOR_INTERNAL  C2    179.250440   58.359815    1.399644 C3   C4   N1  
-ICOOR_INTERNAL  C1    179.835630   60.628601    1.503746 C2   C3   C4  
-ICOOR_INTERNAL  C16     0.250127   61.218574    1.403246 C2   C3   C4  
-ICOOR_INTERNAL  C17  -179.399303   58.784241    1.506067 C16  C2   C3  
-ICOOR_INTERNAL  HC5   120.917576   71.516849    1.097399 C9   N4   C10 
-ICOOR_INTERNAL  HC6  -124.113842   67.833806    1.094153 C9   N4   C10 
-ICOOR_INTERNAL  HC7   117.885715   70.256589    1.096190 C10  C9   O3  
-ICOOR_INTERNAL  HO1    56.085089   73.583094    0.977924 O3   C10  C9  
-ICOOR_INTERNAL  HC8   117.212829   71.866498    1.097442 C11  C10  O4  
-ICOOR_INTERNAL  HO2   -65.475766   71.153891    0.975975 O4   C11  C10 
-ICOOR_INTERNAL  HC9   118.848578   69.761348    1.096695 C12  C11  O5  
-ICOOR_INTERNAL  HO3    42.009787   73.417655    0.982433 O5   C12  C11 
-ICOOR_INTERNAL  HC10  119.675313   67.975885    1.093391 C13  C12  O6  
-ICOOR_INTERNAL  HC11 -118.089241   68.907447    1.094889 C13  C12  O6  
-ICOOR_INTERNAL  HO4   -45.238652   75.401964    0.978978 O6   C13  C12 
-ICOOR_INTERNAL  HN1   150.266633   64.485834    1.009484 N2   C6   C7  
-ICOOR_INTERNAL  HC4  -179.796723   61.752207    1.088977 C3   C4   C2  
-ICOOR_INTERNAL  HC12  179.017223   59.130102    1.086315 C15  C14  C16 
-ICOOR_INTERNAL  HC1   123.942511   69.530161    1.094611 C1   C2   C3  
-ICOOR_INTERNAL  HC2     4.685727   68.445271    1.095335 C1   C2   C3  
-ICOOR_INTERNAL  HC3  -114.889587   69.776769    1.094531 C1   C2   C3  
-ICOOR_INTERNAL  HC13   65.476409   69.834292    1.094564 C17  C16  C2  
-ICOOR_INTERNAL  HC14 -174.870441   68.384989    1.095223 C17  C16  C2  
-ICOOR_INTERNAL  HC15  -55.678700   69.528158    1.094635 C17  C16  C2  
-CHI   1   N4   C9  C10   O3 
-CHI   2   C9  C10  C11   O4 
-CHI   3  C10   C9   N4   C8 
-CHI   4  C10  C11  C12   O5 
-CHI   5  C11  C12  C13   O6 
-CHI   6   C9  C10   O3  HO1 
-PROTON_CHI 6 SAMPLES 3 60 -60 180 EXTRA 0
-CHI   7  C10  C11   O4  HO2 
-PROTON_CHI 7 SAMPLES 3 60 -60 180 EXTRA 0
-CHI   8  C11  C12   O5  HO3 
-PROTON_CHI 8 SAMPLES 3 60 -60 180 EXTRA 0
-CHI   9  C12  C13   O6  HO4 
-PROTON_CHI 9 SAMPLES 3 60 -60 180 EXTRA 0
