+++ tmol/tests/ligand/_parity_helpers.py
@@ -14,7 +14,7 @@ from types import SimpleNamespace
 
 from tmol.ligand import params_file, params_io, prepare_single_ligand
 from tmol.ligand.detect import nonstandard_residue_info_from_smiles_via_mol2
-from tmol.ligand.params_reference import (
+from tmol.tests.ligand.params_reference import (
     StrictComparison,
     compare_params_strict,
     generated_fields_from_preparation,
@@ -62,7 +62,7 @@ def chi_axes_equivalent(prep, ref, *, view=None) -> bool:
     :func:`compare_semantic`, then compare the resulting axis sets. Returns
     ``False`` if no isomorphism exists (e.g. a mutated reference graph).
     """
-    from tmol.ligand.equivalence import _heavy_atom_name_mapping
+    from tmol.tests.ligand.equivalence import _heavy_atom_name_mapping
 
     if view is None:
         view = reference_view_from_params(ref)
@@ -120,7 +120,7 @@ def reference_view_from_params(ref) -> SimpleNamespace:
     """Build a ``LigandPreparation``-like view from a parsed ``.params``.
 
     Suitable as the *reference* argument to
-    :func:`tmol.ligand.equivalence.compare_ligand_preparations`: it exposes
+    :func:`tmol.tests.ligand.equivalence.compare_ligand_preparations`: it exposes
     ``residue_type.atoms`` / ``residue_type.bonds``, ``partial_charges``, and an
     empty ``cartbonded_params`` (cartbonded is skipped for the semantic check).
     """
