+++ tmol/tests/ligand/test_fixture_integrity.py
@@ -6,12 +6,12 @@ from pathlib import Path
 
 import pytest
 
-from tmol.ligand.fixture_integrity import (
+from tmol.tests.ligand.fixture_integrity import (
     FixtureMismatch,
     read_mol2_summary,
     require_paired_fixture,
 )
-from tmol.ligand.params_reference import parse_reference_params
+from tmol.tests.ligand.params_reference import parse_reference_params
 
 _GROUND_TRUTH = Path(__file__).parent.parent / "data" / "ligand_ground_truth"
 _ORPHAN_MOL2 = _GROUND_TRUTH / "ref1.mol2"
