+++ tmol/tests/ligand/test_smiles_semantic.py
@@ -16,8 +16,8 @@ from dataclasses import replace
 
 import pytest
 
-from tmol.ligand.params_reference import compare_semantic, parse_reference_params
-from tmol.ligand.parity_manifest import load_parity_manifest
+from tmol.tests.ligand.params_reference import compare_semantic, parse_reference_params
+from tmol.tests.ligand.parity_manifest import load_parity_manifest
 from tmol.tests.ligand._parity_helpers import (
     chi_axes_equivalent,
     prepare_seed_entry,
